chore(server): rename email service
