+++ server/src/Infrastructure/Services/EmailService/Emails/ResetPasswordEmail.ts
@@ -1,0 +1,10 @@
+import { SimpleEmailBase, type EmailProps } from "./SimpleEmailBase";
+
+export class ResetPasswordMail extends SimpleEmailBase {
+  constructor(props: EmailProps) {
+    super(props);
+
+    this.options.subject = "Reset Password";
+    this.options.html = "<b>reset password</b>";
+  };
+};
